cmd: add base for argp and subcommand helper function
The argument handling is quite hard to make it good when using
`getopt` with help and usage messages. The `argp` library is a
better choice for this task.

As there are several commands which have subcommands, a new helper
function called `parse_subcommand` has been added to simplify the
migration to `argp`.

It'll take care of moving the subcommand to `argv[0]` and updating
`argc` and `argv` accordingly, but also do `argp_parse`.

Signed-off-by: Tan Siewert <tan@siewert.io>