cmd: devmem: rework for argp
This subcommand differs from the others a bit in regards to
sub-subcommands. `read` and `write` are handled in `parse_opt` directly
instead of creating two new subcommands, because the actual handling
for reading and writing is done in ast_ahb_access, where it checks if
argv[0] is the operation to be performed.

Signed-off-by: Tan Siewert <tan@siewert.io>